Inet type moved to pgtypes pkg
